refactor: refactor project and change strukture
