+++ src/components/ErrorBoundary/ErrorBoundary.tsx
@@ -1,13 +1,9 @@
 import type { ReactNode } from 'react';
+import type {
+  ErrorBoundaryProps,
+  ErrorBoundaryState,
+} from '../../utils/interfaces';
 import { Component } from 'react';
-
-interface ErrorBoundaryProps {
-  children: ReactNode;
-}
-
-interface ErrorBoundaryState {
-  hasError: boolean;
-}
 
 export default class ErrorBoundary extends Component<
   ErrorBoundaryProps,
@@ -22,8 +18,8 @@ export default class ErrorBoundary extends Component<
     return { hasError: true };
   }
 
-  public componentDidCatch(error: Error) {
-    console.log('Error', error);
+  componentDidCatch(error: Error) {
+    console.error('Error', error);
   }
 
   render(): ReactNode {
